refactor: 예외 처리 관련 리팩토링
- `ErrorCode`가 Shopping 비즈니스 관련 Error만 관리하도록 명시
- `CommonExceptionHandler`로 이름을 바꾸고, 기존 `Shppping` 관련 예외 처리에 서버 예외 처리도 추가
